requirements.txt
Cython>=0.22
cmdstanpy==0.9.5
pystan~=2.19.1.1
numpy>=1.15.4
pandas>=1.0.4
matplotlib>=2.0.0
LunarCalendar>=0.0.9
convertdate>=2.1.2
holidays>=0.11.3.1
setuptools>=42
setuptools-git>=1.2
python-dateutil>=2.8.0
tqdm>=4.36.1
wheel>=0.37.0
grpcio==1.43.0
beautifulsoup4==4.10.0
google==3.0.0
soupsieve==2.3.1
protobuf==3.19.4
fbprophet==0.7.1